requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
playwright==1.51.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyee==12.1.1
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
